pyproject.toml
[tool.poetry]
name = "orderbook"
version = "0.1.0"
description = "A base implementation of orderbook"
authors = ["mbalashovv <isdedtool@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.113.0"
uvicorn = "^0.20.0"
pydantic = "1.10.16"
dependency-injector = "^4.43.0"
yoyo-migrations = "^9.0.0"
python-dotenv = "^1.0.1"
aiopg = "^1.4.0"
jsf = "0.7.1"

[tool.poetry.group.linters.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
pylint = "^3.3.1"
ruff = "^0.6.9"
pylint-pytest = "2.0.0a0"

[tool.poetry.group.formatters.dependencies]
autoflake = "^2.3.1"
isort = "^5.10.1"
black = "^24.10.0"
docformatter = "^1.7.5"
add-trailing-comma = "^2.2.3"

[tool.poetry.group.sasts.dependencies]
safety = "^2.3.3"
bandit = "^1.8.0"
itsdangerous = "^2.2.0"

[tool.poetry.group.tests.dependencies]
pytest-asyncio = "^0.24.0"
coverage = "^7.6.2"
pytest-cov = "^5.0.0"
pytest-repeat = "^0.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
